refactor: PostTable 관련 UI들 디렉토리 위치 변경
