+++ src/entities/post/ui/LikeDislikeStats.tsx
