+++ src/feature/post/ui/index.ts
@@ -7,3 +7,5 @@ export * from "./PostEditDialog"
 export * from "./Pagination"
 export * from "./PostDetailDialog"
 export * from "./PostAddButton"
+export * from "./PostTable"
+export * from "./PostTableItem"
