--- src/widgets/post/ui/index.ts
@@ -1,3 +1,0 @@
-export * from "./PostTable"
-export * from "./LikeDislikeStats"
-export * from "./PostTableItem"
